Move starting.py
